move profiles and regions handling to commands package
